+++ cmd/backup.go
@@ -86,6 +86,7 @@ func RunBackup(ctx context.Context, flagMap map[string]any) error {
 		preflight.NewValidator(),
 		pathsync.NewPathSyncer(
 			runConfig.Engine.Performance.BufferSizeKB,
+			runConfig.Engine.Performance.ReadAheadLimitKB,
 			runConfig.Engine.Performance.SyncWorkers,
 			runConfig.Engine.Performance.MirrorWorkers,
 		),