+++ cmd/list.go
@@ -65,6 +65,7 @@ func RunList(ctx context.Context, flagMap map[string]any) error {
 		preflight.NewValidator(),
 		pathsync.NewPathSyncer(
 			runConfig.Engine.Performance.BufferSizeKB,
+			runConfig.Engine.Performance.ReadAheadLimitKB,
 			runConfig.Engine.Performance.SyncWorkers,
 			runConfig.Engine.Performance.MirrorWorkers,
 		),