+++ cmd/prune.go
@@ -90,6 +90,7 @@ func RunPrune(ctx context.Context, flagMap map[string]any) error {
 		preflight.NewValidator(),
 		pathsync.NewPathSyncer(
 			runConfig.Engine.Performance.BufferSizeKB,
+			runConfig.Engine.Performance.ReadAheadLimitKB,
 			runConfig.Engine.Performance.SyncWorkers,
 			runConfig.Engine.Performance.MirrorWorkers,
 		),